Refactor SONiC functionality into dedicated sonic.py module
This commit reorganizes the codebase by moving all SONiC-related functionality
into a dedicated module for better code organization and maintainability:

- Created osism/tasks/conductor/sonic.py with all SONiC functions:
  - DEFAULT_SONIC_ROLES constant for allowed device roles
  - Device metadata extraction functions (platform, hostname, MAC address)
  - Port configuration parsing and SONiC config generation
  - Complete sync_sonic implementation with device filtering

- Refactored osism/tasks/conductor/__init__.py:
  - Removed all SONiC helper functions and business logic (170+ lines)
  - Simplified sync_sonic Celery task to delegate to sonic module
  - Cleaned up imports and removed unused dependencies

- Updated osism/tasks/conductor.py imports accordingly

The Celery task interface remains unchanged while significantly improving
code modularity and separation of concerns.
