+++ osism/tasks/conductor.py
@@ -2,11 +2,7 @@
 
 from osism.tasks.conductor import (
     app,
-    get_device_hostname,
-    get_device_mac_address,
-    get_device_platform,
     get_ironic_parameters,
-    get_port_config,
     sync_netbox,
     sync_ironic,
     sync_sonic,
@@ -14,11 +10,7 @@ from osism.tasks.conductor import (
 
 __all__ = [
     "app",
-    "get_device_hostname",
-    "get_device_mac_address",
-    "get_device_platform",
     "get_ironic_parameters",
-    "get_port_config",
     "sync_netbox",
     "sync_ironic",
     "sync_sonic",
